backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
pymysql==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
requests==2.31.0
email-validator==2.1.0
phonenumbers==8.13.27